フォロー機能とフォロー解除機能の正常系テストを作成
